[Bronze V] Title: 이상한 기호, Time: 208 ms, Memory: 17656 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -1,14 +1,14 @@
 import java.util.Scanner;
 
 public class Main {
-    public static int func(int a, int b){
+    public static long func(long a, long b){
         return (a+b)*(a-b);
     }
     public static void main(String[] args) {
         Scanner scan = new Scanner(System.in);
-        int a = scan.nextInt();
-        int b = scan.nextInt();
-        int result = func(a,b);
+        long a = scan.nextInt();
+        long b = scan.nextInt();
+        long result = func(a,b);
 
         System.out.println(result);
     }
